Imported Upstream version 6.0.0~beta

+++ ni/src/ncl/AddIntrinsics.c
@@ -1,7 +1,6 @@
 
-
 /*
- *      $Id: AddIntrinsics.c,v 1.34 2009-12-30 22:11:07 huangwei Exp $
+ *      $Id: AddIntrinsics.c,v 1.35 2010-04-14 21:29:47 huangwei Exp $
  */
 /************************************************************************
 *									*
@@ -560,7 +559,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -578,7 +577,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -587,6 +586,7 @@ void _NclAddIntrinsics
 	args[3].dim_sizes[0] = 1;
 	args[3].n_dims = 1;
 	_NclRegisterFunc(_NclIfbindirread,args,"fbindirread",4,IFUNC);
+
 	args = NclCalloc(4,sizeof(NclArgTemplate));
 	args[0].arg_data_type = _NclLookUp("string");
 	args[0].dim_sizes[0] = 1;
@@ -596,7 +596,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -628,7 +628,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 0;
 	args[2].n_dims = 1;
 	args[3].arg_data_type = _NclLookUp("string");
@@ -657,7 +657,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -671,7 +671,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
